requirements.txt
django==3.2.4
django-extensions==3.1.3
djangorestframework==3.12.4
